Move ApiSelector to a separate file
