+++ components/generic-documentation/src/renderers/helpers/styled.ts
